+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 - Initialize CLI ([#1](https://github.com/AleksanderWWW/go-datareader-cli/pull/1))
+- Add Stooq command and implement runner struct ([#4](https://github.com/AleksanderWWW/go-datareader-cli/pull/4))
 
 ### Changes
 - Handle default date values ([#2](https://github.com/AleksanderWWW/go-datareader-cli/pull/2))